requirements.txt
SQLAlchemy==2.0.23
icecream==2.1.3
pytest==7.4.3
pandas==2.1.3
coverage==7.3.2
python-dotenv==1.0.0